+++ RELEASE-NOTES.md
@@ -12,6 +12,7 @@ Under development
 * Added transit layer support for Google Maps (`layers=transit`) (by @acnetj)
 * Added `egMapsGeoJsonCacheTtl` setting and optional caching for GeoJSON files
 * Fixed display of Maps category on Special:SpecialPages
+* Removed unused `tilt` parameter for Google Maps, including the `egMapsGMaps3DefaultTilt` setting
 
 ## Maps 6.3.0