+++ RELEASE-NOTES.md
@@ -3,6 +3,10 @@ different releases and which versions of PHP and MediaWiki they support, see the
 [platform compatibility tables](INSTALL.md#platform-compatibility-and-release-status).
 
 
+## Maps 7.1.0
+
+Under development
+
 ## Maps 7.0.0
 
 Released on December 16th, 2018.